DAOS-2846 common: enlarge task stack buffer
The DAOS-2846 looks like the update task stack overflowed.
But I cannot find the root point for that. I make a temporary
patch with enlarging the task stack buffer. That will allow
other patches to pass Jenkins tests. It may be not the final
solution, we can replace it when we have more suitable one
in the future.

Signed-off-by: Fan Yong <fan.yong@intel.com>