DAOS-2884 dtx: commit dtx during dtx_resync
The object rebuild logic will trigger DTX resync before scanning.
For performance consideration, the DTX maybe just inside the CoS
cache on the leader. But that is not enough. Because if there is
another failure on current leader (rebuild 32 does) after current
rebuild, then related DTX (that was committable and cached on the
old failed leader) on the new leader will be 'prepared', but the
new target (by the first rebuild just now) does not has such DTX
information. So DTX resync for the second rebuild will be confused.
The right behavior is to force commit the committable DTX during
the DTX resync.

Split vos_dtx_check_committable() into two parts:
vos_dtx_check(): checks the DTX on-dist status.
vos_dtx_lookup_cos(): check whether the DTX in CoS cache or not.

Signed-off-by: Fan Yong <fan.yong@intel.com>