DAOS-[2373,2374,2376] Security: Credential Signing Support (#831)
DAOS-2373: Certificate loading in daos_agent for signing of credentials was
primarily added in earlier patches however this patch adds the passing of the
necessary values into the drpc module handling code for use in signing the
credentials.

DAOS-2374: Server certificate loading in daos_server for credential validation
and validate token

daos_server is modified to load the appropriate agent certificate based on the
origin field of the credential and validates the signature. If the server is
running in insecure mode then it does not attempt to load certificates or
validate them.

DAOS-2376: Modify daos_agent to sign credential and fill in host identifier in
the verifier token and daos_server to validate the credential.

daos_agent is modified to always fill in the origin field with agent for single
certificate support initially. The agent is also modified to sign the credential
with the certificates passed down into the drpc security module instead of just
placing a hash as the verifier.

Signed-off-by: David Quigley <david.quigley@intel.com>