DAOS-2846 object: initialize shard_auxi_args for IO retry (#887)
It is possible that the IO RPC hits some retriable error during
prepare phase (such as -DER_STALE because of stale pool map ver)
for the first time sending. At that time, related shard_auxi_args
has not been initialized yet. So when retries this RPC, we need to
initialize them properly.

Signed-off-by: Fan Yong <fan.yong@intel.com>