DAOS-3035 Allow DAOS RPM builds and/or tests to fail
We occasionally run into times when the DAOS RPM build fails due to
component library incompatibilities. This is a temporary situation which
will be resolved when component libraries are properly maintaining
versions and we have a repository server capable of serving multiple
versions of libraries at the same time.

Until then we need to occasionally disable requiring the RPM builds
and/or tests to pass.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>