DAOS-623 build: Add a CODEOWNERS file (#875)
And populate it initially with release-engineering being added as a
reviewer on PRs touching utils/build.config.

This is to give release-engineering a heads-up when a component sha1 is
going to be updated.

Note that this doesn't require an automatically added reviewer to
review, it just adds them to the list of reviewers. To require reviews
from CODEOWNERS we'd need to enable that feature on our project. Per
https://help.github.com/en/articles/about-code-owners:

  When someone with admin or owner permissions has enabled required
  reviews, they also can optionally require approval from a code owner
  before the author can merge a pull request in the repository.

We don't have that option enabled in this repo.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>