CORCI-261: build: Global product versioning
Add a VERSION file for all things that want to display a version
to be able to read from.

Add an scons target "release" that takes a RELEASE variable to create
a new release from.  This updates:
- the VERSION file
- the RPM daos.spec
and then commits the changes and creates an annotated tag.

Get the version of the libdaos library to import from the VERSION file.

Quick-build: true
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>