DAOS-3108 vos: Cache the DTX state in sorted iterator
For sorted iterator, we should cache the state at the time
it was taken.  This prevents potential to dereference stale
pointers internally.  The user can still mess up by dereferencing
the bio_addr_t but we leave this to aggregation to handle.

Traversing the iterator to pass a deleted cached entry isn't
necessarily a bug and we shouldn't crash if it happens.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>