DAOS-3102 dtx: race between normal IO and rebuild (#924)
Currently, we allow to handle normal IO during rebuild. It
is possible that during the rebuild process, the normal IO
request has created the obj/key (that to be rebuilt by the
rebuild ULT) on the in-rebuilding target before the rebuild
ULT doing that. Then when rebuild ULT tries to rebuild the
same obj/key, related DTX for the normal IO request may be
not ready for commit. So the rebuild ULT may hit 'prepared'
DTX when accesses such obj/key. Under such case, we should
not ask rebuild ULT to retry via returning -DER_INPROGRESS,
instead, the rebuild ULT can (partially) commit the normal
DTX with shared obj/key to make rebuild to go ahead.

Signed-off-by: Fan Yong <fan.yong@intel.com>