DAOS-3126 tests: Fix timeouts for daos_agent and ssh
agent_utils.py attempts to start daos_agent with ssh and this may ocassional
cause errors. If the agent timeout is hit before ssh times out it will appear
that the agent hung when it was ssh. To be able to differentiate between an
agent startup hang and an ssh timeout the timeout for agent startup has been
raised to 15 seconds from 5 seconds and the timeout for ssh connect has been
lowered to 10 seconds from the system default TCP timeout of 20 seconds.

Signed-off-by: David Quigley <david.quigley@intel.com>