DAOS-3181 test: Unhandled error in ftest.sh teardown
There is an unhandled error in ftest.sh where ((i-1)) evaluates to
zero in the case whre i is 1.  This is a normal shell return value but
should be || trued out to avoid the unhandled error.

There is also a logic inversion that is causing the teardown to run all
5 iterations even though the first one is successful.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>