DAOS-3196 debug: key format for debug log
The current rebuild system always assumes key is a reasonable sized
string while generating log, this assumption obviously is not true.

This patch adds a function to truncate large string key to 63 bytes
and converts non-printable key to "????"

This patch also removes/disables a few D_DEBUGs which are not very
useful for debugging. With this small change, size of log files
generated can reduce about 10%.

Signed-off-by: Liang Zhen <liang.zhen@intel.com>