DAOS-3156 mgmt: Remove mode validation from pool create (#961)
While mode is no longer used in pool creation it is still part of the public
API and is checked in the client library. Remove the check for valid mode on
pool create. The public API is left to avoid huge changes in light of the fact
that pool creation will no longer be allowed from the old dmg tool soon.i

Signed-off-by: David Quigley <david.quigley@intel.com>