DAOS-2792 tests: Add yaml file host replacement in launch.py (#746)
Provide a mechanism similar to ftest.sh which replaces the boro-[A-H]
server/client placeholders in the test yaml files with a specified
list of host names to enable easier manual execution of the tests
without needing to modify the yaml file to use the reserved servers.

Also fixes an issue where using explict filenames instead of a tag
resulted in running all tests and the filenames.

Updated launch.py to attempt to attempt to replace placeholders in the
yaml file for each tests specified - not just fail at the first missing
replacement occurance - and run any test that has a valid yaml file.

The full_pool_container_create.yaml change has been verified in as far
as starting daos_server on the host specified in the yal file.  There
are additional issues with this test that prevent it from running and
DAOS-3118 has been created to address them.  The execution of the test
will no longer be included in this PR's test tags as it will cause CI to
fail.

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>