DAOS-2934 control: pull go-spdk package into main repo (#1021)
go-spdk bindings are being maintained in a stand-alone github repo
and used as dependency provided in Go vendor directory.

Whilst this was initially done to enable wider usage of the bindings
it seems practical to bring inside DAOS as the use cases are tightly
coupled to the project. Doing so will reduce maintenance burden.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>