DAOS-3234 api: Update pool_info specification and usage.
because of the pi_bits field pool_info is an input as well as
an output parameter of pool_query()/connect() so mark it as
such.  Update all sites in dfuse/dfs to zero this struct before
passing it in.

Remove the valgrind suppression for the uninitialised read this
was causing.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>