DAOS-3196 vos: force VEA reclaim on cont destroy (#1044)
When a NVMe extent is freed by vea_free(), it won't be available
for next allocation immediately, instead, it'll stay in an aging
buffer for few seconds, so that it could be coalesced with other
recent freed extents, the coalesced extent will be unmapped (NVMe
unmap) and being available for next allocation once it's expired.

On container destroy, there will be huge amount of NVMe extents
being freed all together, so we'd force all of them available for
allocation immediately.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I531ca4a37f92623a9a1d33e7081c531009e8d7d4