DAOS-3101 dtx: sync object
New object based RPC: DAOS_OBJ_RPC_SYNC. When client wants to sync object
against some epoch, it will send DAOS_OBJ_RPC_SYNC RPC to leader replica.
If the object data resides on multiple redundancy groups, the RPC will be
sent to leader replicas in all related redundancy groups.

On the server side, related RPC handler will commit the DTXs that modify the
specified object and which epochs are not newer than the specified epoch. If
some DTX which epoch is older than the given epoch but not ready for commit
at that time, then it will be aborted, related IO RPC handler will reply to
the client -DER_INPROGRESS that will cause client to retry related RPC with
newer epoch sometime later.

If the RPC sponsor does not know which epoch should be used for sync, it can
use DAOS_EPOCH_MAX, then the leader replica will replace it with its current
highest epoch for the sync boundary, all current committable DTXs belonging
to such object will be committed by force, and such epoch will be returned
to the client.

Please note that above sync is object based and persistent (we store related
epoch in the vos object on disk). After the sync done, nobody can modify the
object with the epoch that is not newer than the sync epoch. The sync epoch
is incremental only. If the new sync epoch is not newer than current stored
sync epoch, then it will be ignored.

To avoid confusing, we do not directly export object sync API to application,
instead, the object sync can be internally used by other DAOS components. For
example, the object consistency verification logic needs to sync the object.

Signed-off-by: Fan Yong <fan.yong@intel.com>