Fixed Bug Related to Reselecting Down Targets for Rebuild
Fixed a bug that occurred when selecting rebuild targets for a
layout when most/all targets were down in the original layout.
The bug occurred because the targets that were in the layout
where not then properly discluded from future selections of
fallback targets.

In order to keep track of targets that are unavailable and have
already been considered as fallback targets a d_list_t that contains
all previously considered targets. This is used to build a bitmap
and then ignore those when calling jump consistent hashing in the
same way as before.

Signed-off-by: Peter Fetros <peter.fetros@intel.com>