DAOS-3275 control: rediscover storage during scan
If prepare NVMe is performed while the server is running (from
management tool), scan will not show updated results (devices will
not be shown even though they are accessible to SPDK).

This is because the devices are not rediscovered during the scan
process (results are cached).

Don't cache and rediscover devices on each scan, overhead is deemed
minimal and the benefit of having up-to-date information beneficial.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>