DAOS-2478 duns: use Lustre specifics in libduns (full deps)
This patch allows for libduns concerned methods to become
aware and use of Lustre specific features (foreign LOV/LMV).
This is a 1st shoot, fully (build/install/run) dependent with
Lustre includes/libs, since it uses the associated APIs.
Thus, since our CI is not Lustre aware/ready, build will fail
but it is more for the record, and it will serve as a template
for the next versions which will have no more Lustre
dependencies (dynamic discovery and binding of liblustreapi,
hard-coded re-defines of structs and values, direct ioctl()s
usage, ...).

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>