DAOS-3279 tests: Enhancing the command-utils.py classes.
Implementing a set of new features to the command_utils.py classes
that will be utilized in a future pull requests:
- The yaml namespace is now provided as an input to the
ObjectWthParameters classes - negating the need to override the
get_params() method with the yaml namespace.
- There is now a common run() method for all the classes that inherited
the CommandWithParameters class - now named ExecutableCommand
- FormattedParameter object whose value is a list will now result in a
string where each value in the list is included in the string with the
same argument key.
- Job managers, e.g. orterun, mpirun, srun, etc. are now implemented as
their own ExecutableCommand class with the abilitly to run any other
ExecutableCommand object.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>