DAOS-2855 control: hide access_points from admin user (#1081)
Remove concept of access_points in the management client API and
tools. Assume first host in host list is the access_point for the
moment. Only a single access_point should be specified in server
config file for the moment (and commands will succeed if the
first host in hostlist on client side matches the access_point
specified in server config).

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>