DAOS-3288 Security: Fix certificate checks to allow for stricter permissions (#1114)
The checks for permissions related to certificates enforced specific
permissions. It did now allow for stricter permissions to be allowed on the
files. This patch sets an upper bound on the permissions but allows stricter
permissions. It also fixes the certificate generation script to generate the
files at the appropriate permission level.

Signed-off-by: David Quigley <david.quigley@intel.com>