DAOS-3109 test: D_ASSERTF ignores alternate assert registration (#218)
In order to catch internal assertions in cmocka unit tests,
we need to register mock_assert.   This doesn't work when
the assertion is D_ASSERTF because that code path ignores
the alternate assert function setting.

Also, register mock_assert in unit tests so internal assertions
are properly caught and reported by cmocka.  This doesn't
work for test_linkage.cpp but that is a separate issue that
doesn't need to be solved for this ticket

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>