DAOS-3109 test: Register mock_assert with gurt in unit tests (#1127)
This will enable us to catch internal assertions in the tests
and properly report them in jenkins rather than just exiting
the test application.

Note that this currently only works with D_ASSERT.  Cart
has a pull request (No 218) to fix that issue.  The version
of cart can be updated independently, however.   This patch
immediately solves half of the issue at hand.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>