DAOS-XXX: Make some tweaks to punch model
Use VOS tx flags.   This is a step toward keeping commit information
in a separate table.   From VOS perpsective, it wants to know if data
is committed or if not committed, whether it's uncommitted state is
known or not (e.g. it's known to be uncommitted if current target is
the leader.  The intent can be used at that point to decide what to
do.   This change is to support caching the entire incarnation log
in the object cache in an intent independent manner.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>