DAOS-3115 vos: Cleanup incarnation log a bit
This cleans up ilog_fetch interface, removing the epoch range
argument.   It updates it so that ilog_fetch can cache a
previous value.  For this purpose, a tree version is added.
Removes the object attribute interfaces.  The semantics of
these APIs with punch model change is weird so remove them
for now and rework them if needed in the future.  They are
unused.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>