DAOS-2753 vos: Update punch model for keys (#1056)
The incarnation log (ilog) is introduced.  It's used to log
two things
1. All punches
2. The first update after a punch (or the set of updates that
   may be first once committed).

The fork on punch model for keys is replaced with logging.  This
simplifies the btree used for keys as a key is created and the
incarnation log is scanned to determine existence at a given
point in time.

Iteration is modified in VOS to return future punches for
visible keys to support rebuild.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>