DAOS-3115 vos: Cleanup incarnation log a bit (#1214)
This cleans up ilog_fetch interface, removing the epoch range
argument.   It updates it so that ilog_fetch can cache a
previous value.  For this purpose, a tree version is added.
Removes the object attribute interfaces.  The semantics of
these APIs with punch model change is weird so remove them
for now and rework them if needed in the future.  They are
unused.
Other changes:Python3 build warnings
*Fix an issue in checksum test passing 0 epoch
*Fix a memory leak in key query
*Add valgrind runtime check if valgrind-devel is installed
*Also reduce size of test when full logging is enabled
*Fix a bug in key_query API
*Add ILOG_REMOVED state to deal with AVAILABLE_DIRTY.  For now,
we don't aggregate entries so just ignore such entries.  This
can be improved when I focus on obj/key aggregation

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>