DAOS-3223 tests: Use system UID and GID for good UID and GID
Realized that container1.open(pool_handle2, containerUUID2)
overwrites the members in container1 and it effectively
becomes container2, so we skip this case

Modified the logic so that 2 try-except blocks will cover
all 3 cases and show appropriate failure logs

Updated open method in TestContainer. Checked that only
one test that's not skipped uses this class and verified
that this change doesn't affect. It seems only the tests
use this class and not other helper classes or functions

Signed-off-by: Makito Kano <makito.kano@intel.com>