DAOS-2429 dtx: store committed DTX in SCM without index
When a DTX is committed, it is added into the in-DRAM committed
DTX table that is organized as a B+tree. If related DAOS server
restarts because of some reason, for supporting DAOS server to
rejoin system (if is not evicted), then we need to re-establish
the indexed committed DTX table in DRAM. So we will store the
committed DTX entries in SCM without index (only append) when
it is committed.

A dedicated ULT will scan the non-indexed committed DTX entries
in SCM to re-generate the indexed committed DTX table in DRAM
when open the container.

Signed-off-by: Fan Yong <fan.yong@intel.com>