DAOS-2283 dfuse: Export pool/container handle from dfuse to il.
Add additional ioctls for the interception library to be able
to export the global versions of the pool and container handles
to the interception library.  This should speed up connection
time in the IL and hence improve perforamnce.

The code checks for the uid/gid of the I/O request to avoid
leaking handles across different users, so make the code back
off to the previous method if the new IOCTLs do not work for
any reason.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>