Added/Refactored Check to See if Entire Domain is Used.
Added a check to get target to determine if the entire
bottom level domain has been used and if so, clear it.

This should only happen if the entire pool map is used, so
either an object that is striped across more targets then the
pool map contains, or if the object is striped across more
than (targets - failed targets)

Also refactored the check/clear code for upper level domains.

Signed-off-by: Peter Fetros <peter.fetros@intel.com>