Change pydaos module to not call daos_fini() until after all refs are dropped.
If there is an exception when running then the module was being unloaded
before objects were destroyed, so create a object for the module and
keep a reference to this in container objects.

This correctly handles shutdown in both the python2 and python3 cases,
and when an exception is raised in python3.  When an exception is raised
in python2 it does mean that daos_fini() does not get called but this
is preferable to calling daos_fini() too early which can lead to a
segfault.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>