DAOS-2881 control: Fix race in harness start test
Didn't see this in local testing or CI, but apparently on
some systems the race detector was triggered when running the
new harness test. This commit feeds the outer test values into
the goroutine that starts the harness, and also waits for the
inner goroutine to finish before starting the next loop.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>