DAOS-3592 control: map health stats to ctrlr
There should be a one-to-one mapping between NVMe health stats struct
and NVMe controller. Currently when displaying a single controller
all health stats from all controllers are shown.

When returning health stats from the SPDK bindings, supply a PCI
address as part of the struct enabling a single health stats struct
to be tied to the relevant controller in the go server code.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>