DAOS-3594 control: Remove log calls from server.ext (#1336)
Fixes a nil pointer dereference crash caused by the ext struct
having a nil logger instance.

This regression was introduced by 539f2d0, which did away with
the global logger. It added a log member to the ext struct, but
did not plumb in a way to set that member via the Configuration
constructor. Unfortunately, this was not detected by unit tests
because there is no coverage of the ext struct.

Also removes unused methods on the ext struct, another step along
the way toward removing it entirely.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>