DAOS-2529 api: dRPC handler for list pools
DRAFT/WIP

With this change, protobuf definitions, generated interface code
and a dRPC handler function are added to access the DAOS I/O server
management service list pools functionality. This will enable
the DAOS control plane tools (e.g., dmg utility) to implement a
command to list pools in a specified DAOS system.

This is a companion interface to a previously-implemented
C API and CaRT RPC mechanism to the same behavior.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>