DAOS-1122 tests: fix mixing suite/sub setup/teardown
This patch fixes cases where group/suite and sub-tests
setup/teardown usage is mixed. This implies to use some
rules in order to prevent double free of the test_arg_t
struct which address is passed using the only
void **state argument of each [sub_]setup()/subtest/
[sub_]teardown routines.
If a state is created using a group_setup routine, it is
expected to be freed by the group_teardown routine and
must not be by any sub_setup/subtest/sub_teardown subsets.
If a sub_setup creates a state, it is expected to be
freed by the sub_teardown routine in the subset, and if
a group_state exists already it must be preserved.

Change-Id: Ifae896101da161d48d8757e1745ec57dce53d8aa
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>