DAOS-2529 api: daos_mgmt_list_pools() and dmg list-pools (#1317)
This commit provides an initial portion of a broader pool/container
listing feature being added to the DAOS API and admin utilities.

With this change a new management service API is added to return
the number of pools in the specified system. When a caller optionally
provides buffer space, a list of basic pool information records
are returned. Each record contains for the given pool its UUID
and the list of pool service replica ranks.

The dmg admin utility adds a list-pools command that accepts a
DAOS system name, invokes the API issuing a request to the
DAOS management service, and prints the results (pool list).

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>