DAOS-3632 container: Retrun HAE on cont query
Return the minimal "Highest Aggregated Epoch" among all targets
on container query, that'll be useful for getting an overview of
the container aggregation progress.

This patch also fixed a race of aggregation vs. snapshot deletion:
When the aggregation ULT yield in vos_aggregate(), the HAE could
be reset to zero by snapshot deletion, the intention is to inform
next round of aggregation to start from epoch 0, so that data from
deleted snapshot can be aggregated, however, in this case, the HAE
will be bumped mistakenly when vos_aggregate() resume.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I1d05ad9a623d7d45e90a644a15dc3436fd030f8e