DAOS-2529 api: dRPC handler for list pools
With this change, protobuf definitions, generated interface code
and a dRPC handler function are added to access the DAOS I/O server
management service list pools functionality. This will enable
the DAOS control plane tools (e.g., dmg utility) to implement a
command to list pools in a specified DAOS system.

This is a companion interface to a previously-implemented
C API and CaRT RPC mechanism to the same behavior.

Also with this change a placeholder for unit testing of the dRPC
handler is added to srv_drpc_tests, and that binary is added to
the list of tests invoked by the run_test.sh script.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>