DAOS-2888 tests: Added create_dmg and destroy_dmg in test_utils.py
Tested with most of the tests that use TestPool and DaosPool

Also added get_service_replicas_from_stdout in dmg_utils.py to
extract service replicas from the output string.

host_port parameter for pool_create and pool_destroy is
optional because when we create a pool, passing in these values
will cause an error.

Signed-off-by: Makito Kano <makito.kano@intel.com>