DAOS-3116 vos: Optimize incarnation log update (#1390)
Add a new vos_ilog_update API to check the log first before
calling ilog_update. It will first fetch the existing log and then
check to see if the creation time is later than any punches
(committed or uncommitted).

A combination of COS removal and update only checking
the previous entry means that we can add hundreds of
incarnation log entries on non-leader ranks. This solves
part of the problem by making sure we don't add new
entries if the log has committed creation time.

We need to also add something, perhaps indexed by
ilog offset, force commit of transactions if the incarnation
log is growing. This would be similar to old COS except
commit is not required for correctness, only for avoiding
too many unnecessary log entries.

With full logging on IO9, I was seeing logs with thousands
of modifications (tree_version) on non-leader ranks. With
this change and a much longer run, it cut down to 90 or
so. So we will still need something to give commit hints
for commit on heavily shared objects/keys. Aggregation
will help as well.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>