DAOS-2529 api: pool container listing
With this change, a user-level administration (daos) utility command
'pool list-containers' is added. Also included are a C language API,
supporting CaRT RPC, and IO server pool/container service functions.
Given a pool handle, and a response buffer provided by the client,
return a list of container UUIDs and the total number of containers
created in the pool. If no response buffer is provided, only the
number of containers in the pool is returned. If a buffer is provided
but is too small to store the result, an error -DER_TRUNC is returned
along with the number of containers in the pool.

The core container service function not require a pool
connection/handle. This is done to support a future dRPC interface
providing access to the feature from the administrator management
utility (dmg). However, the CaRT RPC handler included in this change
does enforce a requirement for a pool handle, since this interface
will be engaged by user-level utilities such as daos.

Also with this change, the existing pool list functionality
has been changed to align with the error handling approach of
this container list feature. Also, API arguments for pool and container
listing are now ordered to be consistent with other DAOS APIs.

Signed-off-by: Ken Cain <kenneth.c.cain@intel.com>