DAOS-3627 vos: Return -DER_SHUTDOWN for vos_obj_hold on killed pool (#1394)
A new assertion has been added to vos_iterator fetch functions to
ensure that we don't yield a ULT without a reprobe. This caused
an issue when the pool is dying, however. The aggregation thread
tries to initialize a nested DKEY iterator which is an equivalent
operation to vos_iter_fetch and triggers an assertion when the
key doesn't exist. In this case, vos_obj_hold is called on the
object and it returns -DER_NONEXIST because vp_dying is set.

Change the error code to -DER_SHUTDOWN for now until a better one
can be defined in gurt. This will avoid triggering the assertion
and allow aggregation to shutdown gracefully.

Also, I realized that a secondary fetch of the incarnation log
after ilog_update (in vos_ilog_update) isn't needed as the fetched
log already contains the information needed for update.

Also change evt_delete to use DAOS_INTENT_PURGE
Fix a bug in vos_ilog_fetch. When refreshing a log, we need to reset the error code

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>