DAOS-1492 ARRAY & DFS: Better short read handling
This patch uses the data returned in the sgl for reads when holes
in the array are read. The total number of holes is calculated which
only indicates the possibility of a POSIX short-read.

In DFS, if the DAOS Array short-fetch is detected, we check the size
of the array and compare it against the offset read to check for a short
read. This improves read performance when data read does not contain
any holes in the file by not having to do a get_size for every
dfs_read operation.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>