DAOS-3532 vos: evict obejct from cache if modification aborted
It is possible that the modification has created the object and put
it in the VOS object cache. If the modification is aborted for some
reason (such as failed to modify something), then we should remove
the object from cache, because the cached obejct still references
the new created dkey/dkey/sv/ev records that will be auto released
when related PMDK transaction is aborted.

Signed-off-by: Fan Yong <fan.yong@intel.com>