DAOS-2478 duns: switch to only Lustre build deps mode
As due to licensing it is not possible to copy/use Lustre source
code, change from full (build/install/run) no Lustre deps mechanics
to build deps only. This means that to allow for Lustre specifics
to be enabled, libduns must be built on a platform where
lustre[-client] distro is installed.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: I94dd4b243fe783d08b25e0526cf1a5a41247233a