python: Fix kv issues in python and add daosdbm module.
Fix crashes in some kv operations where the type was not as
expected.
Change the output type to be non-truncated byte strings.
Change the install path in setup.py
Raise KeyError if key not found to match other dict like objects
Add support for del() of keys.
Add daosdbm for named KV stores.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>