DAOS-3755 vos: Assertion failure in key_tree_prepare on fetch (#1438)
If the key is punched before an update, it is possible to have
a key created but not have any actual subtree initialized.
An assertion was triggering in this case for HDF5

Add a regression test as well.

This is new with punch model change as we allow under punch
even in non-rebuild case.   We could disallow this by using
the rebuild flag but I don't see it as being a big deal.  I
imagine MVCC will prevent it in non-rebuild case in the
future.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>