DAOS-3786 vos: dtx_handle::dth_ent can be NULL for vos_dtx_prepared
It is possible that the modification request sponsored by the client
changes nothing on the server. For example, the given IOD for update
is zero size (daos_recx_t::rx_nr is 0), under such case, it will not
generate DTX entry on the server. So vos_dtx_prepared needs to check
such case and return directly.

Signed-off-by: Fan Yong <fan.yong@intel.com>