CART-730 swim: introduce INACTIVE member status (#186)
During bootstrapping not all members initialized in the same time.
So, to avoid false positive DEAD detection during bootstrapping
let's introduce INACTIVE status for member which mean the following:
 - member will be INACTIVE until first ping response or remove
 - each member will try to ping it but just ignore if it fails

Use RPC with reply for INACTIVE pings. So, if node is not
bootstrapped yet but able to reply to RPC it's treated as in
progress. If there is no reply from node treated it as DEAD.

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>