DAOS-723 Build: add compiler hardening to daos build
Release requirements require using available compiler facilities to harden the
built binaries against buffer overflows. This patch adds several desired
hardening flags to the cflags used to build DAOS and its dependencies and also
fixes a resulting bug that was found in the dfuse tests as a result of enabling
these flags.

Signed-off-by: David Quigley <david.quigley@intel.com>