DAOS-3815 vos: init ioc::ic_umoffs for EV record on NVMe
Originally, we did not assign value to ioc::io_umoffs for EV record
on NVMe case. It was not problem before we check the ioc::ic_umoffs
for such case. With patch 8679ec77f42edcfdcd1ab06a6fb7ec05fe8fc65b,
we need to check ioc::ic_umoffs for both SCM and VNMe cases. So set
it as UMOFF_NULL for EV record on NVMe.

Signed-off-by: Fan Yong <fan.yong@intel.com>