DAOS-3839 security: Assert for impossible invalid ACE
In practice we shouldn't be able to get to the situation
where the ace_size is invalid in flatten_aces(). To
allocate the input buffer, we should have already called
get_flattened_ace_size(), which checks the result of
daos_ace_get_size() for each ACE. To change the result
between the two calls, we'd need to experience a memory
corruption or some other alarming disruption of the
internal state of the system.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>