DAOS-3841 vos: On fetch, we are not handling holes properly
evt_find was not returning ea_inob if the extent is empty.
Move the initialization of ea_inob to evt_ent_array_fill so it's
not dependent on in-tree extents.

Without the inob value, vos_obj_fetch was not inserting a biov
entry for the hole causing issues with filling the input sgl.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>