DAOS-3730 tests: Starting agent and server with common params
Expanding upon the recent changes to the functional tests starting of
the daos_agent and daos_server in which yaml configuration files are
utilized, the common parts of those yaml files are now defined in a
single object.  This includes the transport credentials and access
points In addition a single class maintains the lists of all the agents
and servers being run in the test and it provides a single source from
which to obtain data required by the test from the agent or server
configuration files.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>