DAOS-3840 tests: CI tests take advantage of OPA adapters (#1520)
CI functional testing will now use an active ib0 device for test
execution instead of the eth0 device if it is available on the launch
node.

The OFI_INTERFACE environment variable default value will now be set by
launch.py using the list of active devices on the launch node.  The code
that defines the daos_server yaml configuration file will use the
OFI_INTERFACE setting as the default 'fabric_iface' entry.  The ability
to specify a specific interface through the test yaml file is still
retained, as well as the ability to export an OFI_INTERFACE value to
launch.py to override the default.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>