DAOS-3882 object: Use server-side HLC as modification epoch
Currently, we do not have efficient way to guarantee the uniqueness
of client-side HLC among multiple processes of client nodes. If we
use client-side HLC as related modification epoch, then different
modifications may use the same epoch, that will cause fake overlap
in the evtree. So now, we have to use server-side (DTX leader) HLC
as the modification epoch until we have some better solution (that
is necessary for distributed transaction in the future).

Signed-off-by: Fan Yong <fan.yong@intel.com>